correct typo
git-svn-id: https://arducopter.googlecode.com/svn/trunk@450 f9c3cf11-9bcb-44bc-f272-b75c42450872

+++ libraries/APM_ADC/APM_ADC.h
@@ -16,7 +16,7 @@ class APM_ADC_Class
   public:
 	APM_ADC_Class();  // Constructor
 	void Init();
-	float Ch(unsigned char ch_num);     // HJI Changed from int to float
+	int Ch(unsigned char ch_num);     
 };
 
 extern APM_ADC_Class APM_ADC;
